+++ src/components/filter_group/filter_button.js
@@ -22,6 +22,7 @@ export const EuiFilterButton = ({
   color,
   hasActiveFilters,
   numFilters,
+  numActiveFilters,
   isDisabled,
   isSelected,
   type,
@@ -49,13 +50,24 @@ export const EuiFilterButton = ({
     textProps && textProps.className,
   );
 
+  let dataText;
+  if (typeof children === 'string') {
+    dataText = children;
+  }
+
   const buttonContents = (
     <Fragment>
-      <span className="euiFilterButton__textShift" data-text={children}>
+      <span className="euiFilterButton__textShift" data-text={dataText}>
         {children}
       </span>
       {numFiltersDefined &&
-        <EuiNotificationBadge className="euiFilterButton__notification">{numFilters}</EuiNotificationBadge>
+        <EuiNotificationBadge
+          className="euiFilterButton__notification"
+          size="m"
+          color={isDisabled || !hasActiveFilters ? 'subdued' : 'accent'}
+        >
+          {numActiveFilters || numFilters}
+        </EuiNotificationBadge>
       }
     </Fragment>
   );
@@ -91,9 +103,15 @@ EuiFilterButton.propTypes = {
    */
   hasActiveFilters: PropTypes.bool,
   /**
-   * Adds a notification with number
+   * Pass the total number of filters available and it will
+   * add a subdued notification badge showing the number
    */
   numFilters: PropTypes.number,
+  /**
+   * Pass the number of selected filters and it will
+   * add a bright notification badge showing the number
+   */
+  numActiveFilters: PropTypes.number,
   /**
    * Applies a visual state to the button useful when using with a popover.
    */