+++ src/components/filter_group/index.d.ts
@@ -2,7 +2,7 @@ import { CommonProps } from '../common';
 import { IconType, IconSize } from '../icon'
 /// <reference path="../button/index.d.ts" />
 
-import { SFC, ButtonHTMLAttributes, HTMLAttributes } from 'react';
+import { Component, FunctionComponent, ButtonHTMLAttributes, HTMLAttributes } from 'react';
 
 declare module '@elastic/eui' {
   /**
@@ -11,7 +11,16 @@ declare module '@elastic/eui' {
    * @see './filter_button.js'
    */
 
-  export const EuiFilterButton: SFC<EuiButtonEmptyProps>;
+  export interface EuiFilterButtonProps {
+    numFilters?: number;
+    numActiveFilters?: number;
+    isSelected?: boolean;
+    isDisabled?: boolean;
+    type?: string;
+    grow?: boolean;
+    noDivider?: boolean;
+  }
+  export const EuiFilterButton: FunctionComponent<EuiButtonEmptyProps & EuiFilterButtonProps>;
 
   /**
    * Filter group type defs
@@ -19,7 +28,7 @@ declare module '@elastic/eui' {
    * @see './filter_group.js'
    */
 
-  export const EuiFilterGroup: SFC<CommonProps & HTMLAttributes<HTMLDivElement>>;
+  export const EuiFilterGroup: FunctionComponent<CommonProps & HTMLAttributes<HTMLDivElement>>;
 
   /**
    * Filter select item type defs
@@ -32,7 +41,6 @@ declare module '@elastic/eui' {
     checked?: FilterChecked
   }
 
-  export const EuiFilterSelectItem: SFC<CommonProps &
-    ButtonHTMLAttributes<HTMLButtonElement> & EuiFilterSelectItemProps
-  >;
+  export const EuiFilterSelectItem: Component<CommonProps &
+    ButtonHTMLAttributes<HTMLButtonElement> & EuiFilterSelectItemProps>;
 }